refactor(#587): split WorkflowController into 4 partial files by concern
WorkflowController.cs was a 352-line file housing 19 endpoints across four
concerns (definitions, execution, user tasks, instances). The class was
already declared `partial`, signaling original intent. Splitting matches
the documented WorkflowInstance.cs precedent in CLAUDE.md.

Layout:
- WorkflowController.cs (37 lines) — class header with [ApiController] +
  [Route("[controller]")], ctor, fields, and the partial-file index comment.
- WorkflowController.Definitions.cs (94 lines) — deploy, list definitions,
  list instances by key (×2), disable, enable.
- WorkflowController.Execution.cs (90 lines) — start, message, signal,
  evaluate-conditions, complete-activity.
- WorkflowController.UserTasks.cs (156 lines) — tasks list/get + claim/
  unclaim/complete/fail/cancel + the 5 user-task [LoggerMessage] decls.
- WorkflowController.Instances.cs (19 lines) — instance state snapshot.

Invariants preserved:
- All 19 [Http*] route attributes byte-identical → URL surface unchanged
  (verified: pre/post `[Http(Get|Post|...)]` count = 19).
- All 19 [EnableRateLimiting(...)] attributes byte-identical (verified
  count = 19).
- All 5 [LoggerMessage] partial declarations carried into UserTasks.cs.
- [ApiController] + [Route("[controller]")] kept only on the header file;
  other partials declare bare `public partial class WorkflowController`.
- Public consumers of /workflow/* (self-host docker-compose docs + 5
  manual test plans) require zero change.

CLAUDE.md "How to Add a New API Endpoint" updated to point at the
partial-file pattern, mirroring the existing WorkflowInstance bullet.

Closes #587
