+++ src/Fleans/Fleans.Api/Controllers/WorkflowController.cs
@@ -1,13 +1,18 @@
-using System.Text;
 using Fleans.Application;
-using Fleans.Application.QueryModels;
 using Fleans.Infrastructure.Bpmn;
-using Fleans.ServiceDefaults.DTOs;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.RateLimiting;
 
 namespace Fleans.Api.Controllers
 {
+    // WorkflowController is split across partial files by concern:
+    //   WorkflowController.cs              — header: ctor, fields, class-level attributes
+    //   WorkflowController.Definitions.cs  — deploy, list-definitions, list-instances-by-key (×2), disable, enable
+    //   WorkflowController.Execution.cs    — start, message, signal, evaluate-conditions, complete-activity
+    //   WorkflowController.UserTasks.cs    — tasks list/get + claim/unclaim/complete/fail/cancel + LoggerMessage decls
+    //   WorkflowController.Instances.cs    — instance state snapshot
+    // All endpoints live under /workflow/* (preserved exactly). Each partial declares
+    // `public partial class WorkflowController : ControllerBase` without class-level
+    // attributes — the [ApiController] / [Route] below apply class-wide via merge.
     [ApiController]
     [Route("[controller]")]
     public partial class WorkflowController : ControllerBase
@@ -28,325 +33,5 @@ namespace Fleans.Api.Controllers
             _workflowQueryService = workflowQueryService;
             _bpmnConverter = bpmnConverter;
         }
-
-        [EnableRateLimiting("workflow-mutation")]
-        [HttpPost("start", Name = "StartWorkflow")]
-        public async Task<IActionResult> StartWorkflow([FromBody] StartWorkflowRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.WorkflowId))
-            {
-                return BadRequest(new ErrorResponse("WorkflowId is required"));
-            }
-
-            var initialVariables = request.Variables != null
-                ? VariableConverter.ToExpandoObject(request.Variables)
-                : null;
-
-            var instanceId = await _commandService.StartWorkflow(request.WorkflowId, initialVariables);
-            return Ok(new StartWorkflowResponse(instanceId));
-        }
-
-        [EnableRateLimiting("workflow-mutation")]
-        [HttpPost("message", Name = "SendMessage")]
-        public async Task<IActionResult> SendMessage([FromBody] SendMessageRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.MessageName))
-                return BadRequest(new ErrorResponse("MessageName is required"));
-
-            var variables = VariableConverter.ToExpandoObject(request.Variables);
-
-            var result = await _commandService.SendMessage(request.MessageName, request.CorrelationKey, variables);
-
-            if (!result.Delivered)
-                return NotFound(new ErrorResponse(
-                    $"No subscription or start event found for message '{request.MessageName}'"));
-
-            return Ok(new SendMessageResponse(result.Delivered, result.WorkflowInstanceIds));
-        }
-
-        [EnableRateLimiting("workflow-mutation")]
-        [HttpPost("signal", Name = "SendSignal")]
-        public async Task<IActionResult> SendSignal([FromBody] SendSignalRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.SignalName))
-                return BadRequest(new ErrorResponse("SignalName is required"));
-
-            var result = await _commandService.SendSignal(request.SignalName);
-
-            if (result.DeliveredCount == 0 && (result.WorkflowInstanceIds == null || result.WorkflowInstanceIds.Count == 0))
-                return NotFound(new ErrorResponse(
-                    $"No subscription or start event found for signal '{request.SignalName}'"));
-
-            return Ok(new SendSignalResponse(result.DeliveredCount, result.WorkflowInstanceIds));
-        }
-
-        [EnableRateLimiting("workflow-mutation")]
-        [HttpPost("evaluate-conditions", Name = "EvaluateConditions")]
-        public async Task<IActionResult> EvaluateConditions([FromBody] EvaluateConditionsRequest request)
-        {
-            if (request == null)
-                return BadRequest(new ErrorResponse("Request body is required"));
-
-            if (request.Variables == null || request.Variables.Count == 0)
-                return BadRequest(new ErrorResponse("Variables are required for condition evaluation"));
-
-            var variables = VariableConverter.ToExpandoObject(request.Variables);
-            var result = await _commandService.EvaluateConditions(request.WorkflowId, variables);
-            return Ok(new EvaluateConditionsResponse(result.StartedInstanceIds, result.Errors));
-        }
-
-        [EnableRateLimiting("task-operation")]
-        [HttpPost("complete-activity", Name = "CompleteActivity")]
-        public async Task<IActionResult> CompleteActivity([FromBody] CompleteActivityRequest request)
-        {
-            if (request == null || request.WorkflowInstanceId == Guid.Empty)
-                return BadRequest(new ErrorResponse("WorkflowInstanceId is required"));
-            if (string.IsNullOrWhiteSpace(request.ActivityId))
-                return BadRequest(new ErrorResponse("ActivityId is required"));
-
-            var variables = VariableConverter.ToExpandoObject(request.Variables);
-
-            await _commandService.CompleteActivity(request.WorkflowInstanceId, request.ActivityId, variables);
-            return Ok();
-        }
-
-        [EnableRateLimiting("workflow-mutation")]
-        [HttpPost("deploy", Name = "DeployWorkflow")]
-        public async Task<IActionResult> DeployWorkflow([FromBody] DeployBpmnRequest request)
-        {
-            if (string.IsNullOrWhiteSpace(request.BpmnXml))
-                return BadRequest(new ErrorResponse("BpmnXml is required."));
-
-            try
-            {
-                using var stream = new MemoryStream(Encoding.UTF8.GetBytes(request.BpmnXml));
-                var workflow = await _bpmnConverter.ConvertFromXmlAsync(stream);
-                var summary = await _commandService.DeployWorkflow(workflow, request.BpmnXml);
-                return Ok(new DeployBpmnResponse(summary.ProcessDefinitionKey, summary.Version));
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(new ErrorResponse($"Invalid BPMN: {ex.Message}"));
-            }
-        }
-
-        [EnableRateLimiting("read")]
-        [HttpGet("definitions", Name = "ListDefinitions")]
-        public async Task<IActionResult> ListDefinitions(
-            [FromQuery] int page = 1,
-            [FromQuery] int pageSize = 20,
-            [FromQuery] string? sorts = null,
-            [FromQuery] string? filters = null)
-        {
-            var request = new PageRequest(page, pageSize, sorts, filters);
-            var result = await _workflowQueryService.GetAllProcessDefinitions(request);
-            return Ok(result);
-        }
-
-        [EnableRateLimiting("read")]
-        [HttpGet("definitions/{key}/instances", Name = "ListInstancesByKey")]
-        public async Task<IActionResult> ListInstancesByKey(
-            string key,
-            [FromQuery] int page = 1,
-            [FromQuery] int pageSize = 20,
-            [FromQuery] string? sorts = null,
-            [FromQuery] string? filters = null)
-        {
-            var request = new PageRequest(page, pageSize, sorts, filters);
-            var result = await _workflowQueryService.GetInstancesByKey(key, request);
-            return Ok(result);
-        }
-
-        [EnableRateLimiting("read")]
-        [HttpGet("definitions/{key}/{version:int}/instances", Name = "ListInstancesByKeyAndVersion")]
-        public async Task<IActionResult> ListInstancesByKeyAndVersion(
-            string key, int version,
-            [FromQuery] int page = 1,
-            [FromQuery] int pageSize = 20,
-            [FromQuery] string? sorts = null,
-            [FromQuery] string? filters = null)
-        {
-            var request = new PageRequest(page, pageSize, sorts, filters);
-            var result = await _workflowQueryService.GetInstancesByKeyAndVersion(key, version, request);
-            return Ok(result);
-        }
-
-        [EnableRateLimiting("read")]
-        [HttpGet("tasks", Name = "GetPendingTasks")]
-        public async Task<IActionResult> GetPendingTasks(
-            [FromQuery] string? assignee = null,
-            [FromQuery] string? candidateGroup = null,
-            [FromQuery] int page = 1,
-            [FromQuery] int pageSize = 20,
-            [FromQuery] string? sorts = null,
-            [FromQuery] string? filters = null)
-        {
-            var request = new PageRequest(page, pageSize, sorts, filters);
-            var result = await _workflowQueryService.GetPendingUserTasks(assignee, candidateGroup, request);
-            return Ok(result);
-        }
-
-        [EnableRateLimiting("read")]
-        [HttpGet("tasks/{activityInstanceId:guid}", Name = "GetTask")]
-        public async Task<IActionResult> GetTask(Guid activityInstanceId)
-        {
-            var task = await _workflowQueryService.GetUserTask(activityInstanceId);
-            if (task == null)
-                return NotFound(new ErrorResponse($"User task '{activityInstanceId}' not found"));
-
-            return Ok(task);
-        }
-
-        [EnableRateLimiting("task-operation")]
-        [HttpPost("tasks/{activityInstanceId:guid}/claim", Name = "ClaimTask")]
-        public async Task<IActionResult> ClaimTask(Guid activityInstanceId, [FromBody] ClaimTaskRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.UserId))
-                return BadRequest(new ErrorResponse("UserId is required"));
-
-            var task = await _workflowQueryService.GetUserTask(activityInstanceId);
-            if (task == null)
-                return NotFound(new ErrorResponse($"User task '{activityInstanceId}' not found"));
-
-            try
-            {
-                LogUserTaskClaim(activityInstanceId, request.UserId);
-                await _commandService.ClaimUserTask(task.WorkflowInstanceId, activityInstanceId, request.UserId);
-                return Ok();
-            }
-            catch (InvalidOperationException ex)
-            {
-                return Conflict(new ErrorResponse(ex.Message));
-            }
-        }
-
-        [EnableRateLimiting("task-operation")]
-        [HttpPost("tasks/{activityInstanceId:guid}/unclaim", Name = "UnclaimTask")]
-        public async Task<IActionResult> UnclaimTask(Guid activityInstanceId)
-        {
-            var task = await _workflowQueryService.GetUserTask(activityInstanceId);
-            if (task == null)
-                return NotFound(new ErrorResponse($"User task '{activityInstanceId}' not found"));
-
-            LogUserTaskUnclaim(activityInstanceId);
-            await _commandService.UnclaimUserTask(task.WorkflowInstanceId, activityInstanceId);
-            return Ok();
-        }
-
-        [EnableRateLimiting("task-operation")]
-        [HttpPost("tasks/{activityInstanceId:guid}/complete", Name = "CompleteTask")]
-        public async Task<IActionResult> CompleteTask(Guid activityInstanceId, [FromBody] CompleteTaskRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.UserId))
-                return BadRequest(new ErrorResponse("UserId is required"));
-
-            var task = await _workflowQueryService.GetUserTask(activityInstanceId);
-            if (task == null)
-                return NotFound(new ErrorResponse($"User task '{activityInstanceId}' not found"));
-
-            var variables = VariableConverter.ToExpandoObject(request.Variables);
-
-            try
-            {
-                LogUserTaskComplete(activityInstanceId, request.UserId);
-                await _commandService.CompleteUserTask(
-                    task.WorkflowInstanceId, activityInstanceId, request.UserId, variables);
-                return Ok();
-            }
-            catch (InvalidOperationException ex)
-            {
-                return Conflict(new ErrorResponse(ex.Message));
-            }
-        }
-
-        [EnableRateLimiting("task-operation")]
-        [HttpPost("tasks/{activityInstanceId:guid}/fail", Name = "FailTask")]
-        public async Task<IActionResult> FailTask(Guid activityInstanceId, [FromBody] FailTaskRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.ErrorMessage))
-                return BadRequest(new ErrorResponse("ErrorMessage is required"));
-
-            var task = await _workflowQueryService.GetUserTask(activityInstanceId);
-            if (task == null)
-            {
-                if (await _workflowQueryService.UserTaskExists(activityInstanceId))
-                    return Ok();
-                return NotFound(new ErrorResponse($"User task '{activityInstanceId}' not found"));
-            }
-
-            LogUserTaskFail(activityInstanceId, request.ErrorCode);
-            await _commandService.FailUserTask(
-                task.WorkflowInstanceId, activityInstanceId, request.ErrorCode, request.ErrorMessage);
-            return Ok();
-        }
-
-        [EnableRateLimiting("task-operation")]
-        [HttpPost("tasks/{activityInstanceId:guid}/cancel", Name = "CancelTask")]
-        public async Task<IActionResult> CancelTask(Guid activityInstanceId, [FromBody] CancelTaskRequest? request)
-        {
-            var task = await _workflowQueryService.GetUserTask(activityInstanceId);
-            if (task == null)
-            {
-                if (await _workflowQueryService.UserTaskExists(activityInstanceId))
-                    return Ok();
-                return NotFound(new ErrorResponse($"User task '{activityInstanceId}' not found"));
-            }
-
-            LogUserTaskCancel(activityInstanceId);
-            await _commandService.CancelUserTask(task.WorkflowInstanceId, activityInstanceId, request?.Reason);
-            return Ok();
-        }
-
-        [EnableRateLimiting("admin")]
-        [HttpPost("disable", Name = "DisableProcess")]
-        public async Task<IActionResult> DisableProcess([FromBody] ProcessDefinitionKeyRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.ProcessDefinitionKey))
-                return BadRequest(new ErrorResponse("ProcessDefinitionKey is required"));
-
-            var summary = await _commandService.DisableProcess(request.ProcessDefinitionKey);
-            return Ok(summary);
-        }
-
-        [EnableRateLimiting("admin")]
-        [HttpPost("enable", Name = "EnableProcess")]
-        public async Task<IActionResult> EnableProcess([FromBody] ProcessDefinitionKeyRequest request)
-        {
-            if (request == null || string.IsNullOrWhiteSpace(request.ProcessDefinitionKey))
-                return BadRequest(new ErrorResponse("ProcessDefinitionKey is required"));
-
-            var summary = await _commandService.EnableProcess(request.ProcessDefinitionKey);
-            return Ok(summary);
-        }
-
-        [EnableRateLimiting("polling")]
-        [HttpGet("instances/{instanceId:guid}/state", Name = "GetInstanceState")]
-        public async Task<IActionResult> GetInstanceState(Guid instanceId)
-        {
-            var snapshot = await _workflowQueryService.GetStateSnapshot(instanceId);
-            if (snapshot is null)
-                return NotFound(new ErrorResponse($"Instance {instanceId} not found"));
-            return Ok(snapshot);
-        }
-
-        [LoggerMessage(EventId = 8004, Level = LogLevel.Information,
-            Message = "Claiming user task {ActivityInstanceId} for user {UserId}")]
-        private partial void LogUserTaskClaim(Guid activityInstanceId, string userId);
-
-        [LoggerMessage(EventId = 8005, Level = LogLevel.Information,
-            Message = "Unclaiming user task {ActivityInstanceId}")]
-        private partial void LogUserTaskUnclaim(Guid activityInstanceId);
-
-        [LoggerMessage(EventId = 8006, Level = LogLevel.Information,
-            Message = "Completing user task {ActivityInstanceId} by user {UserId}")]
-        private partial void LogUserTaskComplete(Guid activityInstanceId, string userId);
-
-        [LoggerMessage(EventId = 8007, Level = LogLevel.Information,
-            Message = "Failing user task {ActivityInstanceId} with error code {ErrorCode}")]
-        private partial void LogUserTaskFail(Guid activityInstanceId, string errorCode);
-
-        [LoggerMessage(EventId = 8008, Level = LogLevel.Information,
-            Message = "Cancelling user task {ActivityInstanceId}")]
-        private partial void LogUserTaskCancel(Guid activityInstanceId);
     }
 }
